xdr: fix the vecsizer issues
Gluster's protocol was tweaked a lot to handle RDMA of write and read requests.

That too with a opaque type for xdata on these 2 fops, there was a lot
of state machines added while reading data on the wire itself.

With the change of xdr structure, there was changes needed outside of
xlator protocol layer to handle the it properly.

updates #384

Change-Id: I82eef745994873e71b8e6c1de24f2420902222b8
Signed-off-by: Amar Tumballi <amarts@redhat.com>